webapi/tct-geoallow-w3c-tests/geoallow/Geolocation_watchPosition_with_timeout_0_maximumAge_0.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jin,Weihu <weihux.jin@intel.com>

-->
<html>
<head>
  <title>Device Test: watchPosition - should be triggered when timeout is 0 and maximumAge is 0</title>
  <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation" />
  <meta name="assert" content="watchPosition - should be triggered when timeout is 0 and maximumAge is 0" />
  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>
  <script src="testinterface.js"></script>
</head>
<body>
  <div id="log"></div>
  <script>
    var t = async_test(document.title);
    setup({timeout: 20000});

    try {
      if(!check("geolocation", navigator)) {
        t.step(function() {
          assert_unreach("navigator.geolocation missing, test failed");
        });
      } else {
        navigator.geolocation.watchPosition(successCallback, errorCallback, { timeout: 0, maximumAge: 0 });
      }
    } catch(err) {
      t.step(function() {
        assert_unreached("Error message: {" + err.message + " }");
      });
    }

    function successCallback(position) {
      t.done();
    }

    function errorCallback(error) {
      var msg = "";
      if(error.code == 0) {
        msg = "UNKNOW_ERROR";
      } else if (error.code == 1) { // PERMISSION_DENIED
        msg = "The user rejects the browser requests";
      } else if (error.code == 2) { //POSITION_UNAVAILABLE
        msg = "Cannot find your location. Make sure your networkconnection is active and click the link to try again.";
      } else if (error.code == 3) { //TIMEOUT
        t.done();
        return;
      }
      t.step(function () {
        assert_unreached("PositionErrorCallback handle Event is triggered with the exception of "+ msg.toString());
      });
      t.done();
    }
  </script>
</body>
</html>
